falta a funcao clock() colaborar

+++ src/rr.h
@@ -10,7 +10,7 @@
 typedef struct {
 	pid_t pid;
 	int priority; // Quanto menor, mais prioridade
-	double arrive_time;
+	int arrive_time;
 	int* io_times;
 	int* io_types;
 	int io_count;
